ui: backoffice fix assign item api call link
* move function to correct folder
