+++ invenio_app_ils/ui/src/invenio_app_ils/pages/backoffice/LoanDetails/components/AvailableItems/state/actions.js
@@ -1,5 +1,5 @@
 import { IS_LOADING, SUCCESS, HAS_ERROR } from './types';
-import { item as itemApi } from '../../../../../../common/api';
+import { item as itemApi, loan as loanApi } from '../../../../../../common/api';
 import { invenioConfig } from '../../../../../../common/config';
 
 export const fetchAvailableItems = documentPid => {
@@ -16,6 +16,28 @@ export const fetchAvailableItems = documentPid => {
           .withStatus(invenioConfig.items.available.status)
           .qs()
       )
+      .then(response => {
+        dispatch({
+          type: SUCCESS,
+          payload: response.data,
+        });
+      })
+      .catch(error => {
+        dispatch({
+          type: HAS_ERROR,
+          payload: error,
+        });
+      });
+  };
+};
+
+export const assignItemToLoan = (itemId, loanId) => {
+  return async dispatch => {
+    dispatch({
+      type: IS_LOADING,
+    });
+    await loanApi
+      .assignItemToLoan(itemId, loanId)
       .then(response => {
         dispatch({
           type: SUCCESS,
