+++ invenio_app_ils/ui/src/invenio_app_ils/pages/backoffice/LoanDetails/components/LoanMetadata/LoanMetadata.js
@@ -12,7 +12,6 @@ export default class LoanMetadata extends Component {
     this.state = {
       isAvailableItemsVisible: _isEmpty(this.props.loanDetails.item),
     };
-    this.assignItemToLoan = props.assignItemToLoan;
   }
 
   showAvailableItems = (newState = true) => {
